mlp_schedule: refactor common bits to helper functions
