+++ lighthouse/schedule/xegpu/__init__.py
@@ -7,15 +7,29 @@ from .fused_attention_schedule import fused_attention_schedule
 from .xegpu_parameter_selector import XeGPUParameterSelector
 from .matmul_constraints import check_constraints
 from .xegpu_specs import XeGPUSpecs
+from .lowering_common import (
+    bufferize,
+    convert_to_gpu_launch,
+    convert_vector_to_xegpu,
+    outline_gpu_function,
+    vectorize,
+    vectorize_bufferize_and_outline_gpu_func,
+)
 
 __all__ = [
     "XeGPUParameterSelector",
     "XeGPUSpecs",
+    "bufferize",
     "check_constraints",
+    "convert_to_gpu_launch",
+    "convert_vector_to_xegpu",
     "elemwise_schedule",
     "fused_attention_schedule",
     "layer_norm_schedule",
     "mlp_schedule",
+    "outline_gpu_function",
     "softmax_schedule",
+    "vectorize",
+    "vectorize_bufferize_and_outline_gpu_func",
     "xegpu_to_binary",
 ]
